Modularize the codebase into language and debug setup
