+++ package.json
@@ -23,7 +23,8 @@
         "debugger"
     ],
     "activationEvents": [
-        "onLanguage:kotlin"
+        "onLanguage:kotlin",
+        "onDebug"
     ],
     "main": "./out/extension",
     "contributes": {
@@ -57,13 +58,132 @@
                 }
             }
         ],
+        "breakpoints": [
+            {
+                "language": "kotlin"
+            }
+        ],
+        "debuggers": [
+            {
+				"type": "kotlin",
+				"label": "Kotlin",
+				"configurationAttributes": {
+					"launch": {
+						"required": [
+							"projectRoot",
+							"mainClass"
+						],
+						"properties": {
+							"projectRoot": {
+								"type": "string",
+								"description": "Absolute path to your project's root folder",
+								"default": "${workspaceFolder}"
+							},
+							"mainClass": {
+								"type": "string",
+								"description": "The full name of your main class (usually your filename in PascalCase postfixed with Kt)",
+								"default": "YourMainClassKt"
+							},
+							"enableJsonLogging": {
+								"type": "boolean",
+								"description": "Enables logging of debug server JSON messages into a file defined by 'jsonLogFile'",
+								"default": false
+							},
+							"jsonLogFile": {
+								"type": "string",
+								"description": "Enables logging of debug server JSON messages into a file defined by 'jsonLogFile'",
+								"default": "${workspaceFolder}/ktDebugJsonLog.txt"
+							},
+							"logLevel": {
+								"type": "string",
+								"description": "Sets the logging level of the debug server for the messages in the Debug Console [NONE/ERROR/WARN/INFO/DEBUG/TRACE/DEEP_TRACE/ALL]",
+								"default": "INFO"
+							}
+						}
+					},
+					"attach": {
+						"required": [
+							"hostName",
+							"port",
+							"timeout"
+						],
+						"properties": {
+							"hostName": {
+								"type": "string",
+								"description": "The host name of your running JVM",
+								"default": "localhost"
+							},
+							"port": {
+								"type": "number",
+								"description": "The port number of your running JVM",
+								"default": 8000
+							},
+							"timeout": {
+								"type": "number",
+								"description": "The connection timeout",
+								"default": 30000
+							},
+							"enableJsonLogging": {
+								"type": "boolean",
+								"description": "Enables logging of debug server JSON messages into a file defined by 'jsonLogFile'",
+								"default": false
+							},
+							"jsonLogFile": {
+								"type": "string",
+								"description": "Enables logging of debug server JSON messages into a file defined by 'jsonLogFile'",
+								"default": "${workspaceFolder}/ktDebugJsonLog.txt"
+							},
+							"logLevel": {
+								"type": "string",
+								"description": "Sets the logging level of the debug server for the messages in the Debug Console [NONE/ERROR/WARN/INFO/DEBUG/TRACE/DEEP_TRACE/ALL]",
+								"default": "INFO"
+							}
+						}
+					}
+				},
+				"initialConfigurations": [
+					{
+						"type": "kotlin",
+						"request": "launch",
+						"name": "Kotlin Launch",
+						"projectRoot": "${workspaceFolder}",
+						"mainClass": "YourMainClassKt"
+					}
+				],
+				"configurationSnippets": [
+					{
+						"label": "Kotlin: Launch Project",
+						"description": "A new Kotlin launch configuration",
+						"body": {
+							"type": "kotlin",
+							"request": "launch",
+							"name": "Kotlin Launch",
+							"projectRoot": "^\"\\${workspaceFolder}\"",
+							"mainClass": "YourMainClassKt"
+						}
+					},
+					{
+						"label": "Kotlin: Attach VM",
+						"description": "Attaches a debugger to a running JVM",
+						"body": {
+							"type": "kotlin",
+							"request": "attach",
+							"name": "Kotlin Attach",
+							"hostName": "localhost",
+							"port": 8000,
+							"timeout": 30000
+						}
+					}
+				]
+			}
+        ],
         "configuration": {
             "title": "Kotlin",
             "properties": {
                 "kotlin.languageServer.enabled": {
                     "type": "boolean",
                     "default": true,
-                    "description": "Specifies whether the language server should be used. When enabled the extension will provide code completions and linting, otherwise just syntax highlighting. Might require a reload to apply."
+                    "description": "[Recommended] Specifies whether the language server should be used. When enabled the extension will provide code completions and linting, otherwise just syntax highlighting. Might require a reload to apply."
                 },
                 "kotlin.compiler.jvm.target": {
                     "type": "string",
@@ -73,23 +193,28 @@
                 "kotlin.linting.debounceTime": {
                     "type": "integer",
                     "default": 250,
-                    "description": "[DEBUG] Specifies the debounce time limit. Lower to increase responsiveness at the cost of possibile stability issues"
+                    "description": "[DEBUG] Specifies the debounce time limit. Lower to increase responsiveness at the cost of possibile stability issues."
                 },
                 "kotlin.completion.snippets.enabled": {
                     "type": "boolean",
                     "default": true,
-                    "description": "Specifies whether code completion should provide snippets (true) or plain-text items (false)"
+                    "description": "Specifies whether code completion should provide snippets (true) or plain-text items (false)."
+                },
+                "kotlin.debugAdapter.enabled": {
+                    "type": "boolean",
+                    "default": true,
+                    "description": "[Recommended] Specifies whether the debug adapter should be used. When enabled a debugger for Kotlin will be available."
                 },
                 "kotlin.debounceTime": {
                     "type": "integer",
                     "default": 250,
-                    "description": "[DEPRECATED] Specifies the debounce time limit. Lower to increase responsiveness at the cost of possibile stability issues",
+                    "description": "[DEPRECATED] Specifies the debounce time limit. Lower to increase responsiveness at the cost of possibile stability issues.",
                     "deprecationMessage": "Use 'kotlin.linting.debounceTime' instead"
                 },
                 "kotlin.snippetsEnabled": {
                     "type": "boolean",
                     "default": true,
-                    "description": "[DEPRECATED] Specifies whether code completion should provide snippets (true) or plain-text items (false)",
+                    "description": "[DEPRECATED] Specifies whether code completion should provide snippets (true) or plain-text items (false).",
                     "deprecationMessage": "Use 'kotlin.completion.snippets.enabled'"
                 }
             }
@@ -121,6 +246,8 @@
         "request-progress": "^3.0.0",
         "request-promise-native": "^1.0.7",
         "semver": "^6.3.0",
+        "vscode-debugadapter": "^1.35.0",
+        "vscode-debugprotocol": "^1.35.0",
         "vscode-languageclient": "^5.2.1"
     }
 }
